# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.KLoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.KRegisterReq:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - email
    - password
    type: object
  dto.KUpdateUserReq:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      new_password:
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - email
    - new_password
    - password
    type: object
...
  dto.LoginRes:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/dto.User'
    type: object
  dto.RefreshTokenReq:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshTokenRes:
    properties:
      access_token:
        type: string
    type: object
...
  dto.RegisterRes:
    properties:
      user:
        $ref: '#/definitions/dto.User'
    type: object
  dto.ResendVerifyEmailRequest:
    properties:
      email:
        type: string
    type: object
  dto.ResendVerifyPhoneNumberRequest:
    properties:
      phone_number:
        type: string
    type: object
...
  dto.UpdateUserRes:
    properties:
      message:
        type: string
    type: object
  dto.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  dto.VerifyEmailRequest:
    properties:
      email:
        type: string
      verify_code_email:
        type: string
    type: object
  dto.VerifyPhoneNumberRequest:
    properties:
      phone_number:
        type: string
      verify_code_phone_number:
        type: string
    type: object
  dto.VerifyResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8888
info:
  contact:
    email: quangdangfit@gmail.com
    name: Quang Dang
  description: API for user management
  license:
    name: MIT
    url: https://github.com/MartinHeinz/go-project-blueprint/blob/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: User API
  version: "1.0"
paths:
...
  /auth-patient/login:
    post:
      parameters:
      - description: Body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/dto.KLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginRes'
      summary: Login
      tags:
      - users-patient
  /auth-patient/register:
    post:
      parameters:
      - description: Body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/dto.KRegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegisterRes'
      summary: Register new user
      tags:
      - users-patient
  /auth-patient/update-user:
    put:
      parameters:
      - description: Body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/dto.KUpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateUserRes'
      security:
      - ApiKeyAuth: []
      summary: changes the password
      tags:
      - users-patient
  /auth/facebook/callback:
    get:
      parameters:
      - description: Authorization Code from Facebook
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginRes'
      summary: Handles Facebook OAuth2 callback
      tags:
      - users
  /auth/google/callback:
    get:
      consumes:
      - application/json
      description: Logs user in with Google OAuth2
      parameters:
      - description: Authorization code received from Google
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginRes'
      summary: Login with Google
      tags:
      - users
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.User'
      security:
      - ApiKeyAuth: []
      summary: get my profile
      tags:
      - users
  /auth/refresh-token:
    get:
      parameters:
      - description: Body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RefreshTokenRes'
      security:
      - ApiKeyAuth: []
      summary: get my profile
      tags:
      - users
  /auth/resend-verfiy-code-email:
    put:
      parameters:
      - description: Body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/dto.ResendVerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VerifyResponse'
      security:
      - ApiKeyAuth: []
      summary: Verfiy Code for Email
      tags:
      - users
  /auth/resend-verfiy-code-phone-number:
    put:
      parameters:
      - description: Body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/dto.ResendVerifyPhoneNumberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VerifyResponse'
      security:
      - ApiKeyAuth: []
      summary: Verfiy Code for PhoneNumber
      tags:
      - users
  /auth/verfiy-code-email:
    put:
      parameters:
      - description: Body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VerifyResponse'
      security:
      - ApiKeyAuth: []
      summary: Verfiy Code for Email
      tags:
      - users
  /auth/verfiy-code-phone-number:
    put:
      parameters:
      - description: Body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyPhoneNumberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VerifyResponse'
      security:
      - ApiKeyAuth: []
      summary: Verfiy Code for PhoneNumber
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"